Checkstyle added, JavaDocs, code cleanup
