+++ src/main/java/com/epam/lab/paymentsystem/configuration/DispatcherConfiguration.java
@@ -1,5 +1,6 @@
 package com.epam.lab.paymentsystem.configuration;
 
+import java.util.Locale;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -13,7 +14,15 @@ import org.springframework.web.servlet.i18n.CookieLocaleResolver;
 import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
 import org.springframework.web.servlet.view.InternalResourceViewResolver;
 
-import java.util.Locale;
+/**
+ * This class declares three(ТУТ ПОПРАВИТЕ С УЧЕТОМ ВАШИХ БИНОВ) @Bean methods and may be processed
+ * by the Spring container using @Configuration to generate bean definitions and
+ * service requests for those beans at runtime. {@code DispatcherConfiguration} class
+ * configures the following front-end application components: jsp pages, localization.
+ *
+ * @author unascribed
+ * @since 0.0.1
+ */
 
 @Configuration
 @ComponentScan("com.epam.lab.paymentsystem")
@@ -22,6 +31,13 @@ public class DispatcherConfiguration implements WebMvcConfigurer {
   private static final String PREFIX = "/WEB-INF/jsp/";
   private static final String SUFFIX = ".jsp";
 
+  /**
+   * This @Bean method produces {@code ViewResolver} which allows us to set properties
+   * such as prefix or suffix to the view name to generate the final view page URL.
+   *
+   * @return view resolver
+   * @link org.springframework.web.servlet
+   */
   @Bean
   public ViewResolver internalResourceViewResolver() {
     InternalResourceViewResolver bean = new InternalResourceViewResolver();
@@ -30,6 +46,16 @@ public class DispatcherConfiguration implements WebMvcConfigurer {
     return bean;
   }
 
+  /**
+   * This @Bean method produces {@code ResourceBundleMessageSource} which allows to support
+   * internationalization, it requires the capability of resolving
+   * text messages for different locales. Spring’s context is able to resolve text
+   * messages for a target locale by their keys. Typically, the messages for one locale should be
+   * stored in one separate properties file.
+   *
+   * @return Resource Bundle
+   * @link org.springframework.context.support.ResourceBundleMessageSource
+   */
   @Bean
   public ResourceBundleMessageSource messageSource() {
     ResourceBundleMessageSource messageSource = new ResourceBundleMessageSource();
@@ -38,6 +64,13 @@ public class DispatcherConfiguration implements WebMvcConfigurer {
     return messageSource;
   }
 
+  /**
+   * This @Bean method for web-based locale resolution strategies that allows for both locale
+   * resolution via the request and locale modification via request and response.
+   *
+   * @return locale resolver
+   * @link org.springframework.web.servlet
+   */
   @Bean
   public LocaleResolver localeResolver() {
     CookieLocaleResolver localeResolver = new CookieLocaleResolver();
@@ -47,6 +80,13 @@ public class DispatcherConfiguration implements WebMvcConfigurer {
     return localeResolver;
   }
 
+  /**
+   * Add Spring MVC lifecycle interceptors for pre- and post-processing of controller
+   * method invocations. Interceptors can be registered to apply to all requests or be
+   * limited to a subset of URL patterns.
+   *
+   * @param registry interceptor registry
+   */
   @Override
   public void addInterceptors(InterceptorRegistry registry) {
     LocaleChangeInterceptor localeChangeInterceptor = new LocaleChangeInterceptor();
