+++ src/main/java/com/epam/lab/paymentsystem/entities/Role.java
@@ -1,9 +1,19 @@
 package com.epam.lab.paymentsystem.entities;
 
 import com.epam.lab.paymentsystem.entities.enums.Roles;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 
-import javax.persistence.*;
-
+/**
+ * POJO class object, class which is a role database table representation.
+ * This class implements mapping fields with the corresponding cells of the database table columns.
+ */
 @Entity
 @Table(name = "roles")
 public class Role {
